moved language selector to ts
