+++ app/react/I18N/index.js
@@ -1,4 +1,4 @@
-import I18NMenu from './components/I18NMenu';
+import { i18NMenuComponent } from './components/I18NMenu';
 import I18NLink from './components/I18NLink';
 import I18N from './components/I18N';
 import Translate from './components/Translate';
@@ -8,4 +8,14 @@ import I18NUtils from './utils';
 import I18NApi from './I18NApi';
 import * as actions from './actions/I18NActions';
 
-export { I18NMenu, I18NLink, I18NUtils, I18N, Translate, TranslateForm, t, I18NApi, actions };
+export {
+  i18NMenuComponent as I18NMenu,
+  I18NLink,
+  I18NUtils,
+  I18N,
+  Translate,
+  TranslateForm,
+  t,
+  I18NApi,
+  actions,
+};
